+++ plantcv/geospatial/analyze/dsm.py
@@ -1,25 +1,38 @@
 # Analyze Digital Surface Model (DSM) over many regions
 from plantcv.geospatial._helpers import _gather_ids, _show_geojson
-from plantcv.plantcv import outputs, params
+from plantcv.plantcv.classes import Spectral_data
+from plantcv.plantcv import outputs, params, fatal_error
+from plantcv.plantcv.transform import rescale
+from plantcv.plantcv._debug import _debug
+from rasterio.plot import plotting_extent
+from matplotlib import pyplot as plt
 from rasterstats import zonal_stats
 import numpy as np
 import geopandas
 
 
 def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
-    """A function that analyzes elevation averages over regions and outputs data.
-    Inputs:
-    dsm          = Spectral_Data object of geotif data, used for affine metadata
-    geojson      = Path to the shape file containing the regions for analysis
-    lower        = Lower percentile cut off, input as a list formatted  default lower=25
-    upper        = Upper percentile cut off, input as a list formatted  default upper=90
-    label        = Optional label parameter, modifies the variable name of
-                   observations recorded (default = pcv.params.sample_label).
+    """
+    A function that analyzes elevation averages over regions and outputs data.
 
-    :param dsm: [spectral object]
-    :param geojson: str
-    :param percentile: list
-    :param label: str
+    Parameters
+    ----------
+    dsm : Spectral_data
+        Spectral_data object of geotif data, used for affine metadata
+    geojson : str
+        Path to the shape file containing the regions for analysis
+    lower : int, optional
+        Lower percentile cut off, input as a list formatted  default lower=25
+    upper : int, optional
+        Upper percentile cut off, input as a list formatted  default upper=90
+    label : str, list, optional
+        Optional label parameter, modifies the variable name of observations
+        recorded (default = pcv.params.sample_label).
+
+    Returns
+    -------
+    bounds : list
+        List of geojson bounds analyzed
     """
     # DSM tifs contain just one band of data, so make the array 2D
     dsm_data = dsm.array_data[:, :, 0]
@@ -98,3 +111,60 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
                                  cmap='viridis', vmin=min_elevation, vmax=max_elevation)
 
     return plotting_img
+
+
+def height_subtraction(dsm1, dsm0):
+    """
+    A function that subtracts the height of one DSM from the height of another and outputs a Spectral_data object instance.
+
+    Parameters
+    ----------
+    dsm1 : Spectral_data
+        Spectral_data object of geotif DSM data - DSM with plant height
+    dsm0 : Spectral_data
+        Spectral_data object of geotif DSM data - DSM of bare ground
+
+    Returns
+    -------
+    subtracted_dsm : Spectral_data
+        New Spectral_data object with dsm1 - dsm0
+    """
+    # Check the coordinate reference system (CRS) is the same for both of the DSMs
+    if dsm1.metadata["crs"] != dsm0.metadata["crs"]:
+        fatal_error("The two input DSMs do not have the same coordinate reference system (CRS).")
+
+    # DSM tifs contain just one band of data, so make the array 2D
+    dsm1_data = dsm1.array_data[:, :, 0]
+    dsm0_data = dsm0.array_data[:, :, 0]
+
+    # Check the shapes are equivalent
+    if (dsm1_data.shape == dsm0_data.shape) is False:
+        fatal_error("Input DSMs do not have same shape, can be changed with PCV 'resize' function.")
+
+    # Perform the subtraction
+    final_data = dsm1_data - dsm0_data
+    # Scale visualization
+    final_vis = np.nan_to_num(final_data, nan=0.0)
+    debug = params.debug
+    params.debug = None
+    final_vis = rescale(final_vis, min_value=0, max_value=255)
+    params.debug = debug
+
+    # Convert to uint8
+    pseudo_rgb = final_vis.astype(np.uint8)
+
+    # Make a Spectral_data instance before calculating a pseudo-rgb
+    spectral_array = Spectral_data(array_data=final_data,
+                                   max_wavelength=0,
+                                   min_wavelength=0,
+                                   max_value=np.max(final_vis), min_value=np.min(final_vis),
+                                   d_type=np.float32,
+                                   wavelength_dict=None, samples=int(np.shape(final_vis)[1]),
+                                   lines=int(np.shape(final_vis)[0]), interleave=None,
+                                   wavelength_units="nm", array_type="datacube",
+                                   pseudo_rgb=pseudo_rgb, filename=None,
+                                   default_bands=None,
+                                   metadata=dsm0.metadata)
+
+    _debug(visual=pseudo_rgb, filename=os.path.join(params.debug_outdir, f"{params.device}_substracted_dsm.png"))
+    return spectral_array